Moved and renamed _reproject_blocked to reproject.common
